requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
django-rest-swagger
drf-yasg
pyaml
requests 
django-cors-headers
pexels-api
django-storages
numpy
onnxruntime
opencv-python